merge Develop

+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/WxMpMassMessageService.java
@@ -7,7 +7,7 @@ import me.chanjar.weixin.mp.bean.result.WxMpMassUploadResult;
 
 /**
  * <pre>
- * 群发消息服务类
+ * 群发消息服务类.
  * Created by Binary Wang on 2017-8-16.
  * </pre>
  *
@@ -15,33 +15,33 @@ import me.chanjar.weixin.mp.bean.result.WxMpMassUploadResult;
  */
 public interface WxMpMassMessageService {
   /**
-   * 上传群发用的图文消息
+   * 上传群发用的图文消息.
    */
   String MEDIA_UPLOAD_NEWS_URL = "https://api.weixin.qq.com/cgi-bin/media/uploadnews";
   /**
-   * 上传群发用的视频
+   * 上传群发用的视频.
    */
   String MEDIA_UPLOAD_VIDEO_URL = "https://api.weixin.qq.com/cgi-bin/media/uploadvideo";
   /**
-   * 分组群发消息
+   * 分组群发消息.
    */
   String MESSAGE_MASS_SENDALL_URL = "https://api.weixin.qq.com/cgi-bin/message/mass/sendall";
   /**
-   * 按openId列表群发消息
+   * 按openId列表群发消息.
    */
   String MESSAGE_MASS_SEND_URL = "https://api.weixin.qq.com/cgi-bin/message/mass/send";
   /**
-   * 群发消息预览接口
+   * 群发消息预览接口.
    */
   String MESSAGE_MASS_PREVIEW_URL = "https://api.weixin.qq.com/cgi-bin/message/mass/preview";
   /**
-   * 删除群发接口
+   * 删除群发接口.
    */
   String MESSAGE_MASS_DELETE_URL = "https://api.weixin.qq.com/cgi-bin/message/mass/delete";
 
   /**
    * <pre>
-   * 上传群发用的图文消息，上传后才能群发图文消息
+   * 上传群发用的图文消息，上传后才能群发图文消息.
    *
    * 详情请见: http://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421140549&token=&lang=zh_CN
    * </pre>
@@ -53,7 +53,7 @@ public interface WxMpMassMessageService {
 
   /**
    * <pre>
-   * 上传群发用的视频，上传后才能群发视频消息
+   * 上传群发用的视频，上传后才能群发视频消息.
    * 详情请见: http://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421140549&token=&lang=zh_CN
    * </pre>
    *
@@ -64,7 +64,7 @@ public interface WxMpMassMessageService {
 
   /**
    * <pre>
-   * 分组群发消息
+   * 分组群发消息.
    * 如果发送图文消息，必须先使用 {@link #massNewsUpload(WxMpMassNews)} 获得media_id，然后再发送
    * 如果发送视频消息，必须先使用 {@link #massVideoUpload(WxMpMassVideo)} 获得media_id，然后再发送
    * 详情请见: http://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421140549&token=&lang=zh_CN
@@ -74,7 +74,7 @@ public interface WxMpMassMessageService {
 
   /**
    * <pre>
-   * 按openId列表群发消息
+   * 按openId列表群发消息.
    * 如果发送图文消息，必须先使用 {@link #massNewsUpload(WxMpMassNews)} 获得media_id，然后再发送
    * 如果发送视频消息，必须先使用 {@link #massVideoUpload(WxMpMassVideo)} 获得media_id，然后再发送
    * 详情请见: http://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421140549&token=&lang=zh_CN
@@ -84,7 +84,7 @@ public interface WxMpMassMessageService {
 
   /**
    * <pre>
-   * 群发消息预览接口
+   * 群发消息预览接口.
    * 开发者可通过该接口发送消息给指定用户，在手机端查看消息的样式和排版。为了满足第三方平台开发者的需求，在保留对openID预览能力的同时，增加了对指定微信号发送预览的能力，但该能力每日调用次数有限制（100次），请勿滥用。
    * 接口调用请求说明
    *  http请求方式: POST
@@ -98,7 +98,7 @@ public interface WxMpMassMessageService {
 
   /**
    * <pre>
-   * 删除群发
+   * 删除群发.
    * 群发之后，随时可以通过该接口删除群发。
    * 请注意：
    * 1、只有已经发送成功的消息才能删除
@@ -114,6 +114,6 @@ public interface WxMpMassMessageService {
    * @param msgId        发送出去的消息ID
    * @param articleIndex 要删除的文章在图文消息中的位置，第一篇编号为1，该字段不填或填0会删除全部文章
    */
-  void delete(Integer msgId, Integer articleIndex) throws WxErrorException;
+  void delete(Long msgId, Integer articleIndex) throws WxErrorException;
 
 }
