+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/bean/membercard/WxMpMemberCardUpdateResult.java
@@ -28,7 +28,7 @@ public class WxMpMemberCardUpdateResult implements Serializable {
 
   private Integer resultBonus;
 
-  private Integer resultBalance;
+  private Double resultBalance;
 
   @Override
   public String toString() {
